+++ src/components/Badge/Badge.jsx
@@ -5,7 +5,52 @@ import PropTypes from "prop-types";
 // @material-ui/core components
 import withStyles from "@material-ui/core/styles/withStyles";
 
-import badgeStyle from "assets/jss/material-kit-react/components/badgeStyle";
+import {
+  primaryColor,
+  warningColor,
+  dangerColor,
+  successColor,
+  infoColor,
+  roseColor
+} from "assets/jss/material-kit-react";
+
+const badgeStyle = {
+  badge: {
+    marginRight: "3px",
+    borderRadius: "12px",
+    padding: "5px 12px",
+    textTransform: "uppercase",
+    fontSize: "10px",
+    fontWeight: "500",
+    lineHeight: "1",
+    color: "#fff",
+    textAlign: "center",
+    whiteSpace: "nowrap",
+    verticalAlign: "baseline",
+    display: "inline-block"
+  },
+  primary: {
+    backgroundColor: primaryColor
+  },
+  warning: {
+    backgroundColor: warningColor
+  },
+  danger: {
+    backgroundColor: dangerColor
+  },
+  success: {
+    backgroundColor: successColor
+  },
+  info: {
+    backgroundColor: infoColor
+  },
+  rose: {
+    backgroundColor: roseColor
+  },
+  gray: {
+    backgroundColor: "#6c757d"
+  }
+};
 
 function Badge({ ...props }) {
   const { classes, color, children } = props;
