+++ src/components/Card/CardFooter.jsx
@@ -8,7 +8,14 @@ import withStyles from "@material-ui/core/styles/withStyles";
 // @material-ui/icons
 
 // core components
-import cardFooterStyle from "assets/jss/material-kit-react/components/cardFooterStyle";
+const cardFooterStyle = {
+  cardFooter: {
+    display: "flex",
+    alignItems: "center",
+    backgroundColor: "transparent",
+    padding: "0.9375rem 1.875rem"
+  }
+};
 
 function CardFooter({ ...props }) {
   const { classes, className, children, ...rest } = props;
