+++ src/components/Card/CardHeader.jsx
@@ -8,7 +8,34 @@ import withStyles from "@material-ui/core/styles/withStyles";
 // @material-ui/icons
 
 // core components
-import cardHeaderStyle from "assets/jss/material-kit-react/components/cardHeaderStyle";
+import {
+  warningCardHeader,
+  successCardHeader,
+  dangerCardHeader,
+  infoCardHeader,
+  primaryCardHeader
+} from "assets/jss/material-kit-react";
+
+const cardHeaderStyle = {
+  cardHeader: {
+    borderRadius: "3px",
+    padding: "1rem 15px",
+    marginLeft: "15px",
+    marginRight: "15px",
+    marginTop: "-30px",
+    border: "0",
+    marginBottom: "0"
+  },
+  cardHeaderPlain: {
+    marginLeft: "0px",
+    marginRight: "0px"
+  },
+  warningCardHeader,
+  successCardHeader,
+  dangerCardHeader,
+  infoCardHeader,
+  primaryCardHeader
+};
 
 function CardHeader({ ...props }) {
   const { classes, className, children, color, plain, ...rest } = props;
