+++ src/components/Header/HeaderLinks.jsx
@@ -6,9 +6,30 @@ import { Link } from 'react-router-dom';
 import withStyles from "@material-ui/core/styles/withStyles";
 import List from "@material-ui/core/List";
 import ListItem from "@material-ui/core/ListItem";
-// import Button from "components/CustomButtons/Button.jsx";
-// import headerLinksStyle from "assets/jss/material-kit-react/components/headerLinksStyle.jsx";
-import headerStyle from "assets/jss/material-kit-react/components/headerStyle.jsx";
+
+const headerLinkStyle = {
+    inlineBlock: {
+        display: "inline-block",
+        padding: "0px",
+        width: "auto"
+    },
+    block: {
+        color: "inherit",
+        padding: "0.9375rem",
+        fontWeight: "500",
+        fontSize: "12px",
+        textTransform: "uppercase",
+        borderRadius: "3px",
+        textDecoration: "none",
+        position: "relative",
+        display: "block"
+    },
+    list: {
+        marginBottom: "0",
+        padding: "0",
+        marginTop: "0"
+    },
+}
 
 
 function HeaderLinks ({...props}){
@@ -49,4 +70,4 @@ function HeaderLinks ({...props}){
     )
 };
 
-export default withStyles(headerStyle)(HeaderLinks);
+export default withStyles(headerLinkStyle)(HeaderLinks);
